remove on-demand browserify transform
