+++ .gitignore
@@ -6,3 +6,4 @@ test/test-bundle.js
 npm-debug.log
 testdb*
 test/custom-api/worker-bundle.js
+lib/workerified/index.js
