+++ .jshintignore
@@ -1,0 +1,1 @@
+lib/workerified/index.js
