+++ bin/build.js
@@ -1,0 +1,43 @@
+#!/usr/bin/env node
+
+'use strict';
+
+var browserify = require('browserify');
+var UglifyJS = require("uglify-js");
+var Promise = require('bluebird');
+var fs = require('fs');
+var writeFile = Promise.promisify(fs.writeFile.bind(fs));
+var mkdirp = Promise.promisify(require('mkdirp'));
+var derequire = require('derequire');
+
+function browserifyIt(entry) {
+  return new Promise(function (resolve, reject) {
+    var data = '';
+    var b = browserify().plugin('bundle-collapser/plugin');
+
+    b.add(entry);
+    var bundle = b.bundle();
+    bundle.on('data', function (buf) {
+      data += buf;
+    });
+    bundle.on('end', function () {
+      data = derequire(data);
+      data = UglifyJS.minify(data, {
+        fromString: true,
+        mangle: true,
+        compress: true
+      }).code;
+      resolve(data);
+    }).on('error', reject);
+  });
+}
+return mkdirp('./lib/workerified').then(function () {
+  return browserifyIt('./lib/worker/index.js');
+}).then(function (data) {
+  var code = "// this code is automatically generated by bin/build.js\n" +
+    "module.exports = " + JSON.stringify(data) + ';';
+  return writeFile('./lib/workerified/index.js', code, 'utf-8');
+}).catch(function (err) {
+  console.error(err.stack);
+  process.exit(1);
+});
