+++ lib/client/index.js
@@ -6,13 +6,14 @@ var extend = require('js-extend').extend;
 var WorkerPouchCore = require('./core');
 var createWorker = require('./create-worker');
 var isSupportedBrowser = require('./is-supported-browser');
+var workerCode = require('../workerified');
 
 function WorkerPouch(opts, callback) {
 
   var worker = window.__pouchdb_global_worker; // cache so there's only one
   if (!worker) {
     try {
-      worker = createWorker(require('../worker'));
+      worker = createWorker(workerCode);
       worker.addEventListener('error', function (e) {
         if ('console' in global && 'warn' in console) {
           console.warn('worker threw an error', e.error);
