+++ package.json
@@ -21,18 +21,19 @@
     "url": "https://github.com/nolanlawson/worker-pouch/issues"
   },
   "scripts": {
+    "prepublish": "npm run build",
+    "build-workerified-code": "node bin/build.js",
     "test-browser": "./bin/test-browser.js",
     "jshint": "jshint -c .jshintrc lib test/test.js",
     "test": "npm run jshint && npm run test-browser",
-    "build": "mkdirp dist && npm run browserify && npm run min",
+    "build": "mkdirp dist && run-scripts build-workerified-code browserify min",
     "browserify": "browserify -s workerPouch -p bundle-collapser/plugin . | ./bin/es3ify.js | derequire > dist/pouchdb.worker-pouch.js",
     "min": "UGLIFY=1 browserify -s workerPouch -p bundle-collapser/plugin . | ./bin/es3ify.js | derequire | uglifyjs -mc > dist/pouchdb.worker-pouch.min.js",
-    "dev": "browserify test/test.js > test/test-bundle.js && npm run dev-server",
+    "dev": "npm run build-workerified-code && browserify test/test.js > test/test-bundle.js && npm run dev-server",
     "dev-server": "./bin/dev-server.js",
-    "coverage": "npm test --coverage && istanbul check-coverage --lines 100 --function 100 --statements 100 --branches 100",
     "build-custom-worker": "browserify test/custom-api/worker.js --debug > test/custom-api/worker-bundle.js",
-    "test-custom-local": "npm run build-custom-worker && zuul --ui mocha-bdd --local 9000 --no-coverage test/custom-api/test.js",
-    "test-custom": "npm run build-custom-worker && zuul --ui mocha-bdd --phantom --no-coverage test/custom-api/test.js"
+    "test-custom-local": "run-scripts build-workerified-code build-custom-worker && zuul --ui mocha-bdd --local 9000 --no-coverage test/custom-api/test.js",
+    "test-custom": "run-scripts build-workerified-code build-custom-worker && zuul --ui mocha-bdd --phantom --no-coverage test/custom-api/test.js"
   },
   "dependencies": {
     "argsarray": "0.0.1",
@@ -42,9 +43,7 @@
     "js-extend": "^1.0.1",
     "lie": "^2.6.0",
     "pouchdb": "^5.0.0",
-    "pouchdb-binary-util": "^1.0.0",
-    "through2": "^2.0.0",
-    "uglifyify": "^3.0.1"
+    "pouchdb-binary-util": "^1.0.0"
   },
   "devDependencies": {
     "bluebird": "^2.9.24",
@@ -57,24 +56,19 @@
     "es3ify": "^0.1.3",
     "es5-shim": "^4.1.1",
     "http-server": "^0.8.5",
-    "istanbul": "^0.2.7",
     "jshint": "2.8.0",
     "mkdirp": "^0.5.0",
     "mocha": "^2.3.3",
     "phantomjs-prebuilt": "^2.1.7",
     "pouchdb-http-proxy": "^0.10.4",
     "request": "^2.36.0",
+    "run-scripts": "^0.4.0",
     "sauce-connect-launcher": "^0.11.1",
     "selenium-standalone": "^4.7.0",
     "uglify-js": "^2.4.13",
     "watchify": "^3.1.0",
     "wd": "^0.2.21",
     "zuul": "^3.10.1"
-  },
-  "browserify": {
-    "transform": [
-      "./lib/workerify"
-    ]
   },
   "files": [
     "lib",
